+++ src/components/molecules/vTaskItemStudent.unit.js
@@ -146,4 +146,19 @@ describe("@components/molecules/vTaskItemStudent", () => {
 			expect(validator(task)).toBe(false);
 		});
 	});
+
+	it("should display topic", () => {
+		const wrapper = mount(vTaskItemStudent, {
+			...createComponentMocks({
+				i18n: true,
+				vuetify: true,
+			}),
+			vuetify,
+			propsData: {
+				task: openTasksWithDueDate[0],
+			},
+		});
+
+		expect(wrapper.text()).toContain("Thema Malen nach Zahlen");
+	});
 });